Resend 2fa case impl

+++ src/main/java/com/greenapp/authservice/repositories/PlainUserRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.math.BigInteger;
 
 @Repository
-public interface PlainUserRepository extends JpaRepository<PlainUser, BigInteger> {
+public interface PlainUserRepository extends JpaRepository<PlainUser, Long> {
     PlainUser findByMailAddress(String mailAddress);
     boolean existsPlainUserByMailAddress(String mailAddress);
 }
